Move toJson() method To JsonSerializeTrait
